Implemented a basic version of battery widget
BatteryManager interface was not defined in Typescript (probably it was
not widely supported). So I wrote mine. Also `getBattery` method was not
found in navigator object. For Typescript I bind it to the navigator with
a proper return type.

For now each time, we want to create a battery widget we use
`window.createBatteryWidget` method. It returns to an HTML element with
a custom `battery-widget` tag. For now it only shows the level. But with
the complete access to the BatteryManager nearly all functionality is now
in our reach.